John conroy/move files into components (#1041)
* Move files components into new files directory

* Prettier write package.json

* Update remaining file paths

* Add changelog file
