+++ context/app/static/js/components/files/FileBrowser/FileBrowser.spec.js
@@ -3,8 +3,8 @@ import React from 'react';
 import userEvent from '@testing-library/user-event';
 import { render, screen } from 'test-utils/functions';
 
+import DetailContext from 'js/components/Detail/context';
 import FileBrowser from './FileBrowser';
-import DetailContext from '../context';
 import FilesContext from '../Files/context';
 
 const fakeOpenDUA = jest.fn();
