+++ context/app/static/js/components/files/FileBrowser/index.js
