+++ context/app/static/js/components/files/FileBrowser/style.js
