+++ context/app/static/js/components/files/FileBrowserFile/index.js
