+++ context/app/static/js/components/files/FileBrowserFile/style.js
