+++ context/app/static/js/components/files/FileBrowserNode/index.js
