+++ context/app/static/js/components/files/FileBrowserNode/style.js
